Opted to use the term 'Decorator' over 'Modifier'
