--- UITweaks/PanelModifiers/MultiplayerPositionPanelModifier.cs
@@ -1,24 +1,0 @@
-﻿using UITweaks.Models;
-using Zenject;
-
-namespace UITweaks.PanelModifiers
-{
-    public class MultiplayerPositionPanelModifier : PanelModifierBase
-    {
-
-        [Inject] protected override void Init()
-        {
-            base.Init();
-        }
-
-        protected override void ModPanel()
-        {
-            base.ModPanel();
-        }
-
-        protected override void OnDestroy()
-        {
-            base.OnDestroy();
-        }
-    }
-}
